chore: moved lidarr types to model file
